[Add] Archer Firing
[ADD} Blue Arrow Projectile
[ADD] Red Arrow Projectile
[ADD] Proper firing with Cannon Controller
[REMOVE] cube red from camera object
[MODIFY] Movement - is now not possible when firing a shot

BUGS:
- Arrow not properly arrowlike on trajectory. Will likely use actual
physics instead to set trajectory angle on every aspect of the flight

Notes:
- Both arrow objects are comprised of two Rigidbody2Ds so as to create a
drag like effect of arrows. Although this hasn't worked properly and may
need to be readjusted in the future. As a result, CannonController has
to read child object 0 of every arrow to ensure that every rigidbody's
gravity is off when arrow is spawned. This may not be desired and a
proper course of action will be determined in the foreseeable future.