[Add] Wind Effects [Modify] Scripts
[Add]
Added GameObject with environment tag to use WindSpawner script.
Wind only affects projectiles in flight.
In this version, Archer's Blue Shot will be affected by wind; whereas
the Archer's Scarlet Shot will not be affected by wind. The usage of
this is to show the difference of wind effects.
Consequently, the arrow prefabs were modified as a result.
Added Projectile Controller script and this will be modified to codify
projectile effects in the future

[Modify]
Arrow Rotator script is cleaned up and is now using Mathf from Unity
rather than System.