Modified Movement
[Add] Snow Two Prefab. Thinking of ways how to create more deformable
ground. Might remove this if not viable in the current build
[Add] Player Controller. The old movement script Movement is not exactly
suitable, with Freefall Controller. This script is added to complement
the newer archer prefab which is a bit cleaner.
[Add] Archerino Prefab. This is still archer, but I named it like so
because Archerino. The old prefab is a bit clunky. Maybe this one may
not be so clunky at all?
[Modify] CannonController script to use the PlayerController script
instead of the old Movement script
[Modiy] SpawnSnowStage script. This effectively replaces Terrain,
although this interpretation is a bit wasteful so more efficient ways to
spawn the map may be desired
[Remove] Terrain objects from scene.

Notes and tl:dr;s
Still keeping the old movement script for guide, although it may be
slated for deletion.
Leaning towards a pixelated map.
Cleaner movement with no freefal controller