[ADD] Jumping, Orientation Checker
[MODIFY] Player Controller Script - Jumping, by using Right CTRL key.
- Please note that jumping cannot happen while tank is on movement.
- The idea is that choosing to jump constitutes a tactical choice that
should prove to be erroneous or great.
- Plus, the idea is that jumping would require some fuel usage.
- Note that there is some degree of  left/right movement that can be
done on flight.

[ADD] Orientation Checker script
- Detects the floor configuration when tank is on freefall
- Would align the tank accordingly to the floor. This is done by taking
the raycast points from the wheels, then computing the Vector difference
to account for the tank alignment.
- May be clunky and feels unnatural.
- Jumping would naturally make the tank realign with the floor it is
expecting.

[ADD} Floor
- Repurposed Terrain sprite into a floor sprite. This will be used in
the future to detect falling deaths.

[MODIFY] Spawn Snow Stage Script
- Remodeled map

[MODIFY] Deformable Ground Prefabs
- Added "Terrain" tag

[REMOVE] Freefall Controller and Movement scripts
- These scripts were primitive and were conceptualized when maps were
not made. Their jobs are now handled by other scripts.

[MODIFY] Archerino is now Archer
- The new prefab now does more work than its predecessor and would be
used as a model to introduce other tanks.

[TODO]
- PlayerController - readd solution to allow tank to go up inclines. A
cheaper alternative to clear sharp crevices may be desired as jumping is
slated to use fuel
- SpawnSnowStage - a unity stage prefab may be desired instead. This
might be designed by hand instead using unity. Programatically spawning
objects proves to be a strenuous task to the computer and seems to be
better avoided.
- Prefabs for other tanks
- Explosion sprite animations