[Mod] Camera Controller Script
[Added] Object Tracker
1) Object Tracer allow the Camera to follow projectile
1.1) Object Tracer have different modes that affects how camera is
tracing a projectile. For now there are only ZoomIn and ZoomOut effects.
[Add/Modify] Camera Controller Script
1) The Camera now has 2 modes, which are Focus and Free.
1.1) Focus mode allow the Camera to follow the player. It also enable
ObjectTracker
1.2) Free mode allow the player to move Camera when mouse hits screen
boundary. For example, moving mouse all the way up will cause the Camera
to move up. This allow the player to see the entire map. The player can
also use Mouse Wheel to zoom in and zoom out.