[Add] Scorch [Modify] All prefabs, scripts
[Add] Scorch
- Also added Shots 1 and 2, Ember and Phlogiston
- Modified the way Scorch is firing., and also edited the game design
document It now fires from the crystal in a 90 degree angle. This means,
the angle will be changed to -45 to 45 with respect to the positive X
axis

[Modify] Scripts
- Cleaned up a lot of scripts, and removed get and set variables from
most of them (and set them to public). This way, it's easier to think
when coding.
- Also vastly modified the way ArrowRotator script works, making it far
simpler. This change was realized because of the way Scorch has to
rotate.
- Also noticed a minor issue where the projectile was inverted when
facing the other way.

[Modify] Prefabs
- Tweaked a few values inside some of the tanks to ensure that angles of
the shot are respected
- Changed the name of Scarlet Arrow to Triple Strafe, and changed the
name of Blue Arrow to Single Shot (to reflect GDD names)
- Transferred some of the sprites inside some folders. This will be
cleaned in a future update to ensure cleaner release.