Two-Player Update
[Add] Input Manager
-this script latches on to Environmental Variables game object to set
controls depending on player
- Player 1 uses: WASD, space, left ctrl, tab
- Player 2 uses: updownleftright, numpad0, rightctrl, rightshift
- Please note that changes on the shot are not reflected on the UI yet,
and will be worked on once turns have been fixed

[Modify] main.unity
- temporarily turned off Camera Controller to be able to see whole map
- also temporarily turned off Enemy Spawner
- set Archer as Player 1 and Pirate as Player 2. Using controls, feel
free to play with it!

[Modify] Controls
- the way controls are called are now changed in order to gear towards a
local multiplayer setup in the future

Notes:
Changes to the Camera script may be imminent in order to reflect the
local multiplayer style of the game. To work on: HP values of tanks,
Fuel implementation and Turns.